Wrote html for the 3 things we outlined during club
Spent a couple minutes writting some of html for an outline of the navbar which would include the home button. the mentoring section, and the members list. NOTE: this was just to get a feeling of how full the navbar would feel like with the ideas we had discussed during club.